Refactor entity out of player
